requirements.txt
fastapi
uvicorn
psycopg2-binary
httpx